moved lambda_function to app dir for development
